requirements.txt
streamlit==1.37.1
sentence_transformers==3.0.1
chromadb==0.5.11
pdfplumber==0.11.4
python-docx==1.1.2
tiktoken==0.7.0
nltk==3.8.1
python-dotenv==1.0.1
backoff==2.2.1
scikit-learn==1.4.0
langchain==0.1.0
langchain-community==0.0.16
psutil==5.9.8
tqdm==4.66.1
pandas==2.2.0
openpyxl==3.1.2
requests==2.31.0
langchain-core==0.1.19
torch==2.1.2